+++ src/components/index.js
@@ -1,37 +1,26 @@
-// Atoms
-export { Input } from './atoms/Input';
-export { Button } from './atoms/Button';
-export { AddedFile } from './atoms/AddedFile';
-export { ImageInput } from './atoms/ImageInput';
-
-// Organisms
-export { GoodsName } from './organisms/GoodsName';
-export { FileList } from './organisms/FileList';
-
-// Templates
-export { Category } from './templates/Category';
-export { Item } from './templates/Item';
-export { ItemWithTwoCol } from './templates/ItemWithTwoCol/index';
-
-// Templates/Item
-export { ImageAdder } from './templates/Item/ImageAdder';
 // atoms
-export { AppendButton } from 'components/atoms/AppendButton';
+export { Input } from 'components/atoms/Input';
 export { Button } from 'components/atoms/Button';
+export { AddedFile } from 'components/atoms/AddedFile';
+export { ImageInput } from 'components/atoms/ImageInput';
+export { AppendButton } from 'components/atoms/AppendButton';
 export { ButtonRadio } from 'components/atoms/ButtonRadio';
 export { DeleteButton } from 'components/atoms/DeleteButton';
-export { ImageUploader } from 'components/atoms/ImageUploader';
-export { Input } from './atoms/Input';
 export { InputDate } from 'components/atoms/InputDate';
 
+// molecules
+export { ImageUploader } from 'components/molecules/ImageUploader';
+export { ButtonRadioGroup } from 'components/molecules/ButtonRadioGroup';
+export { InputDatePeriod } from 'components/molecules/InputDatePeriod';
+export { ProductOption } from 'components/molecules/ProductOption';
+export { FileList } from 'components/molecules/FileList';
+
 // organisms
-export { ButtonRadioGroup } from 'components/organisms/ButtonRadioGroup';
+export { ImageAdder } from 'components/organisms/ImageAdder';
 export { GoodsName } from 'components/organisms/GoodsName';
-export { InputDatePeriod } from 'components/organisms/InputDatePeriod';
-export { ProductOption } from 'components/organisms/ProductOption';
+export { SalesExposedPeriod } from 'components/organisms/SalesExposedPeriod';
 
 // templates
 export { Category } from 'components/templates/Category';
 export { Item } from 'components/templates/Item';
-export { ItemWithTwoCol } from './templates/ItemWithTwoCol/index';
-export { SalesExposedPeriod } from 'components/templates/SalesExposedPeriod';
+export { ItemWithTwoCol } from 'components/templates/ItemWithTwoCol';
