+++ src/components/organisms/ImageAdder/index.jsx
