+++ src/components/organisms/ImageAdder/style.jsx
