+++ src/components/organisms/SalesExposedPeriod/index.jsx
